added data to console

+++ WeatherService/WeatherData.cs
@@ -30,17 +30,17 @@ namespace WeatherService
 
     public class Main
     {
-        public int temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
-        public int temp_min { get; set; }
-        public int temp_max { get; set; }
+        public float temp { get; set; }
+        public float pressure { get; set; }
+        public float humidity { get; set; }
+        public float temp_min { get; set; }
+        public float temp_max { get; set; }
     }
 
     public class Wind
     {
         public float speed { get; set; }
-        public int deg { get; set; }
+        public float deg { get; set; }
     }
 
     public class Clouds
